Remove unnecessary type casting
The compiler now imports these as their canonical type, allowing us to
directly use without the explicit type cast.

+++ Sources/_TestingInternals/include/Stubs.h
@@ -118,7 +118,7 @@ SWT_DEFINE_ATOMIC_OPERATIONS(void const *_Nullable)
 ///
 /// This function is provided because `MAKELANGID()` is a complex macro and
 /// cannot be imported directly into Swift.
-static LANGID swt_MAKELANGID(int p, int s) {
+static LANGID swt_MAKELANGID(WORD p, WORD s) {
   return MAKELANGID(p, s);
 }
